requirements.txt
Django==4.1.2
openpyxl==3.0.10
django-extensions==3.2.1
python-dateutil==2.8.2
pandas==1.5.1